+++ src/plugins/mpi/pmix/Makefile.am
@@ -34,7 +34,7 @@ if HAVE_PMIX_V1
 
 pkglib_v1dir=$(pkglibdir)
 pkglib_v1_LTLIBRARIES = mpi_pmix_v1.la
-mpi_pmix_v1_la_SOURCES = $(pmix_src)
+mpi_pmix_v1_la_SOURCES = $(pmix_src) pmixp_client_v1.c
 mpi_pmix_v1_la_LIBADD = $(pmix_libadd)
 mpi_pmix_v1_la_LDFLAGS = $(pmix_ldflags) $(PMIX_V1_LDFLAGS)
 mpi_pmix_v1_la_CPPFLAGS = $(AM_CPPFLAGS) $(PMIX_V1_CPPFLAGS) -DHAVE_PMIX_VER=1
@@ -45,7 +45,7 @@ if HAVE_PMIX_V2
 
 pkglib_v2dir=$(pkglibdir)
 pkglib_v2_LTLIBRARIES = mpi_pmix_v2.la
-mpi_pmix_v2_la_SOURCES = $(pmix_src)
+mpi_pmix_v2_la_SOURCES = $(pmix_src) pmixp_client_v2.c
 mpi_pmix_v2_la_LIBADD = $(pmix_libadd)
 mpi_pmix_v2_la_LDFLAGS = $(pmix_ldflags) $(PMIX_V2_LDFLAGS)
 mpi_pmix_v2_la_CPPFLAGS = $(AM_CPPFLAGS) $(PMIX_V2_CPPFLAGS) -DHAVE_PMIX_VER=2
