+++ src/plugins/mpi/pmix/pmixp_client.h
@@ -40,13 +40,51 @@
 
 #include "pmixp_common.h"
 
-extern pmix_server_module_t _slurm_pmix_callbacks;
+#define PMIXP_ALLOC_KEY(kvp, key_str)				\
+{								\
+	char *key = key_str;					\
+	kvp = (pmix_info_t *)xmalloc(sizeof(pmix_info_t));	\
+	(void)strncpy(kvp->key, key, PMIX_MAX_KEYLEN);		\
+}
+
+#define PMIXP_INFO_ADD(kvp, key_str, field, val) {			\
+	int key_num = 0;						\
+	char *key = key_str;						\
+	if (!kvp) {							\
+		kvp = (pmix_info_t *)xmalloc(sizeof(pmix_info_t));	\
+	} else {							\
+		key_num = xsize(kvp) / sizeof(pmix_info_t);		\
+		kvp = (pmix_info_t *)xrealloc(kvp, (key_num + 1) *	\
+			sizeof(pmix_info_t));				\
+	}								\
+	(void)strncpy(kvp[key_num].key, key, PMIX_MAX_KEYLEN);		\
+	PMIX_VAL_SET(&kvp[key_num].value, field, val);			\
+}
+
+#define PMIXP_INFO_SIZE(kvp) (xsize(kvp) / sizeof(pmix_info_t))
+
+#define PMIXP_FREE_KEY(kvp)				\
+{							\
+	xfree(kvp);					\
+}
 
 int pmixp_libpmix_init(void);
 int pmixp_libpmix_finalize(void);
 int pmixp_libpmix_job_set(void);
 void pmix_libpmix_task_set(int rank, char ***env);
-
 void pmix_client_new_conn(int fd);
+
+int pmixp_lib_init(void);
+int pmixp_lib_finalize(void);
+int pmixp_lib_setup_fork(uint32_t rank, const char *nspace, char ***env);
+int pmixp_lib_dmodex_request(pmixp_proc_t *proc, void *dmdx_fn, void *caddy);
+void pmixp_lib_modex_invoke(void *mdx_fn, int status,
+			  const char *data, size_t ndata, void *cbdata,
+			  void *rel_fn, void *rel_data);
+void pmixp_lib_release_invoke(void *rel_fn, void *rel_data);
+int pmixp_lib_is_wildcard(uint32_t rank);
+int pmixp_lib_is_undef(uint32_t rank);
+uint32_t pmixp_lib_get_wildcard(void);
+uint32_t pmixp_lib_get_version(void);
 
 #endif /* PMIXP_CLIENT_H */
