+++ src/plugins/mpi/pmix/pmixp_dmdx.h
@@ -44,7 +44,7 @@
 int pmixp_dmdx_init(void);
 int pmixp_dmdx_finalize(void);
 int pmixp_dmdx_get(const char *nspace, int rank,
-		   pmix_modex_cbfunc_t cbfunc, void *cbdata);
+		   void *cbfunc, void *cbdata);
 void pmixp_dmdx_process(Buf buf, int nodeid, uint32_t seq);
 void pmixp_dmdx_timeout_cleanup(void);
