+++ src/plugins/mpi/pmix/pmixp_nspaces.h
@@ -49,16 +49,11 @@ typedef struct {
 } pmixp_blob_t;
 
 typedef struct {
-	pmix_modex_data_t data;
-	pmix_scope_t scope;
-} pmixp_modex_t;
-
-typedef struct {
 #ifndef NDEBUG
 #define PMIXP_NSPACE_MAGIC 0xCAFED00D
 	int magic;
 #endif
-	char name[PMIX_MAX_NSLEN];
+	char name[PMIXP_MAX_NSLEN];
 	uint32_t nnodes; /* number of nodes in this namespace */
 	int node_id; /* relative position of this node in this step */
 	uint32_t ntasks; /* total number of tasks in this namespace */
@@ -93,7 +88,7 @@ static inline hostlist_t pmixp_nspace_hostlist(pmixp_namespace_t *nsptr)
 }
 
 hostlist_t pmixp_nspace_rankhosts(pmixp_namespace_t *nsptr,
-				  const int *ranks, size_t nranks);
+				  const uint32_t *ranks, size_t nranks);
 int pmixp_nspace_resolve(const char *name, int rank);
 
 size_t pmixp_nspace_mdx_lsize(List l);
