+++ src/plugins/mpi/pmix/pmixp_server.c
@@ -49,8 +49,6 @@
 #include "pmixp_conn.h"
 #include "pmixp_dconn.h"
 
-#include <pmix_server.h>
-
 #define PMIXP_DEBUG_SERVER 1
 
 /*
@@ -722,7 +720,7 @@ static void _process_server_request(pmixp_base_hdr_t *hdr, Buf buf)
 	case PMIXP_MSG_FAN_IN:
 	case PMIXP_MSG_FAN_OUT: {
 		pmixp_coll_t *coll;
-		pmix_proc_t *procs = NULL;
+		pmixp_proc_t *procs = NULL;
 		size_t nprocs = 0;
 		pmixp_coll_type_t type = 0;
 		int c_nodeid;
@@ -1598,17 +1596,17 @@ bool pmixp_server_want_cperf()
 	return _pmixp_cperf_on;
 }
 
-static void _pmixp_cperf_cbfunc(pmix_status_t status,
+static void _pmixp_cperf_cbfunc(int status,
 				const char *data, size_t ndata,
 				void *cbdata,
-				pmix_release_cbfunc_t r_fn,
+				void *r_fn,
 				void *r_cbdata)
 {
 	/* small violation - we kinow what is the type of release
 	 * data and will use that knowledge to avoid the deadlock
 	 */
 	pmixp_coll_t *coll = pmixp_coll_from_cbdata(r_cbdata);
-	xassert(PMIX_SUCCESS == status);
+	xassert(SLURM_SUCCESS == status);
 
 	/*
 	 * we will be called with mutex locked.
@@ -1618,7 +1616,7 @@ static void _pmixp_cperf_cbfunc(pmix_status_t status,
 	slurm_mutex_unlock(&coll->lock);
 
 	/* invoke the callbak */
-	r_fn(r_cbdata);
+	pmixp_lib_release_invoke(r_fn, r_cbdata);
 
 	/* lock it back before proceed */
 	slurm_mutex_lock(&coll->lock);
@@ -1631,11 +1629,11 @@ static void _pmixp_cperf_cbfunc(pmix_status_t status,
 static int _pmixp_server_cperf_iter(char *data, int ndata)
 {
 	pmixp_coll_t *coll;
-	pmix_proc_t procs;
+	pmixp_proc_t procs;
 	int cur_count = _pmixp_server_cperf_count();
 
-	strncpy(procs.nspace, pmixp_info_namespace(), PMIX_MAX_NSLEN);
-	procs.rank = PMIX_RANK_WILDCARD;
+	strncpy(procs.nspace, pmixp_info_namespace(), PMIXP_MAX_NSLEN);
+	procs.rank = pmixp_lib_get_wildcard();
 
 	coll = pmixp_state_coll_get(PMIXP_COLL_TYPE_FENCE, &procs, 1);
 	xassert(!pmixp_coll_contrib_local(coll, data, ndata,
